requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
pathlib2>=2.3.7
azure-ai-agents
azure-ai-projects
azure-identity
python-dotenv
fastapi
starlette
httpx
pytest
pytest-asyncio
pydantic
fastapi
starlette
httpx
pytest
pytest-asyncio